2.8.8

+++ nbd-server.c
@@ -669,11 +669,13 @@ int mainloop(CLIENT *client) {
 
 		if (request.type==NBD_CMD_DISC) {
 			msg2(LOG_INFO, "Disconnect request received.");
-			if (client->difmap) g_free(client->difmap) ;
-                	if (client->difffile>=0) { 
-                		close(client->difffile);
-				unlink(client->difffilename);
-				free(client->difffilename);
+			if(client->server->flags & F_COPYONWRITE) {
+				if (client->difmap) g_free(client->difmap) ;
+                		if (client->difffile>=0) { 
+                			close(client->difffile);
+					unlink(client->difffilename);
+					free(client->difffilename);
+				}
 			}
 			go_on=FALSE;
 			continue;
@@ -977,10 +979,10 @@ int serveloop(SERVER* serve) {
 	int max_fd = serve->socket;
 	fd_set read_fds;
 
-	FD_ZERO(&read_fds);
-	FD_SET(serve->socket, &read_fds);
-
 	for(;;) {
+		FD_ZERO(&read_fds);
+		FD_SET(serve->socket, &read_fds);
+
 		DEBUG("select, ");
 		/* use to select to tell us when a connection is ready to be
 		 * accepted */
